requirements.txt
Django==3.2.11
djangorestframework==3.13.1
numpy==1.21.5
opencv-python
Pillow
tensorflow
